+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ---
 
+#### 29.06.2018
+
+* [Rust] Code Royale - Struct setters and getters (CodinGame)
+
 #### 28.06.2018
 
 * [Rust] Code Royale - Queen struct functionality (CodinGame)